+++ probables/exceptions.py
@@ -6,7 +6,7 @@ class ProbablesBaseException(Exception):
     ''' Base ProbablesBaseException
 
         Args:
-            message (str): The error messge to be reported
+            message (str): The error message to be reported
     '''
     def __init__(self, message):
         self.message = message
@@ -34,8 +34,19 @@ class NotSupportedError(ProbablesBaseException):
     ''' Not Supported Functionality Exception
 
         Args:
-            message (str): The error messge to be reported
+            message (str): The error message to be reported
     '''
     def __init__(self, message):
         self.message = message
         super(NotSupportedError, self).__init__(self.message)
+
+
+class CuckooFilterFullError(ProbablesBaseException):
+    ''' Cuckoo Filter Full Exception
+
+        Args:
+            message (str): The error message to be reported
+    '''
+    def __init__(self, message):
+        self.message = message
+        super(CuckooFilterFullError, self).__init__(self.message)