+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PyProbables Changelog
 
+### Version 0.1.0:
+* Cuckoo Filter
+    * Added basic Cuckoo Filter code
+
 ### Version 0.0.8:
 * Counting Bloom Filter
     * Estimate unique elements added